requirements.txt
tk
mss
numpy
fpdf
matplotlib
Pillow
opencv-python
pymupdf
pyinstaller